catalog-service/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14,<4.0
mysqlclient>=2.2,<3.0
